# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pakistani Job Aggregator API
 * OpenAPI spec version: 0.1.0
 */
export interface CvFileInput {
  file: Blob;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface UserPreferences {
  sectors?: string[];
  locations?: string[];
  experienceLevel?: string;
  minMatchScore?: number;
}

export interface Profile {
  id: string;
  userId: string;
  /** @nullable */
  cvText?: string | null;
  /** @nullable */
  cvUpdatedAt?: string | null;
  preferences?: UserPreferences;
  createdAt?: string;
  updatedAt?: string;
}

export interface ProfileUpdate {
  preferences?: UserPreferences;
}

export interface CvUploadResult {
  success: boolean;
  cvTextLength: number;
  cvUpdatedAt: string;
}

export interface CvSuggestions {
  suggestions: string[];
  generatedAt: string;
}

export interface Job {
  id: string;
  source: string;
  title: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  requirements?: string | null;
  /** @nullable */
  experienceLevel?: string | null;
  /** @nullable */
  jobType?: string | null;
  /** @nullable */
  salaryRange?: string | null;
  /** @nullable */
  postedDate?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  applyUrl?: string | null;
  applicationSteps?: string[];
  /** @nullable */
  matchScore?: number | null;
  createdAt: string;
}

export interface JobListResponse {
  jobs: Job[];
  total: number;
  page: number;
  limit: number;
}

export interface MatchedJob {
  job: Job;
  matchScore: number;
}

export interface MatchedJobsResponse {
  jobs: MatchedJob[];
  hasCV: boolean;
}

export interface SectorStat {
  sector: string;
  count: number;
}

export interface JobStats {
  total: number;
  bySector: SectorStat[];
  recentCount: number;
}

export interface MatchAnalysis {
  jobId: string;
  matchScore: number;
  strengths: string[];
  gaps: string[];
  suggestions: string[];
  cachedAt: string;
}

export type SavedJobStatus = typeof SavedJobStatus[keyof typeof SavedJobStatus];


export const SavedJobStatus = {
  saved: 'saved',
  applied: 'applied',
  interview: 'interview',
  rejected: 'rejected',
  offer: 'offer',
} as const;

export interface SavedJob {
  id: string;
  userId: string;
  jobId: string;
  job?: Job;
  status: SavedJobStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  appliedAt?: string | null;
  createdAt: string;
}

export type SavedJobInputStatus = typeof SavedJobInputStatus[keyof typeof SavedJobInputStatus];


export const SavedJobInputStatus = {
  saved: 'saved',
  applied: 'applied',
  interview: 'interview',
  rejected: 'rejected',
  offer: 'offer',
} as const;

export interface SavedJobInput {
  jobId: string;
  status?: SavedJobInputStatus;
  notes?: string;
}

export type SavedJobUpdateStatus = typeof SavedJobUpdateStatus[keyof typeof SavedJobUpdateStatus];


export const SavedJobUpdateStatus = {
  saved: 'saved',
  applied: 'applied',
  interview: 'interview',
  rejected: 'rejected',
  offer: 'offer',
} as const;

export interface SavedJobUpdate {
  status?: SavedJobUpdateStatus;
  notes?: string;
  appliedAt?: string;
}

export interface SyncResult {
  inserted: number;
  updated: number;
  deleted: number;
  errors: number;
  message: string;
}

export type ListJobsParams = {
sector?: string;
category?: string;
location?: string;
experienceLevel?: string;
jobType?: string;
/**
 * today | week | month
 */
postedWithin?: string;
/**
 * 3days | week | month
 */
deadlineWithin?: string;
minMatchScore?: number;
search?: string;
page?: number;
limit?: number;
};

export type GetMatchedJobsParams = {
limit?: number;
};