# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pakistani Job Aggregator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs with filters
      parameters:
        - name: sector
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: experienceLevel
          in: query
          schema:
            type: string
        - name: jobType
          in: query
          schema:
            type: string
        - name: postedWithin
          in: query
          description: today | week | month
          schema:
            type: string
        - name: deadlineWithin
          in: query
          description: 3days | week | month
          schema:
            type: string
        - name: minMatchScore
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
  /jobs/matched:
    get:
      operationId: getMatchedJobs
      tags: [jobs]
      summary: Get top AI-matched jobs for the current user's CV
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Matched jobs with scores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchedJobsResponse"
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/analyze:
    get:
      operationId: analyzeJobMatch
      tags: [jobs]
      summary: Get AI match analysis for a job vs current user's CV (cached)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchAnalysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /cron/sync-adzuna:
    post:
      operationId: syncAdzuna
      tags: [cron]
      summary: Trigger Adzuna job sync (protected by CRON_SECRET)
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Job:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        title:
          type: string
        company:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        experienceLevel:
          type: ["string", "null"]
        jobType:
          type: ["string", "null"]
        salaryRange:
          type: ["string", "null"]
        postedDate:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        applyUrl:
          type: ["string", "null"]
        applicationSteps:
          type: array
          items:
            type: string
        matchScore:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - source
        - title
        - createdAt
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - jobs
        - total
        - page
        - limit
    MatchedJob:
      type: object
      properties:
        job:
          $ref: "#/components/schemas/Job"
        matchScore:
          type: number
      required:
        - job
        - matchScore
    MatchedJobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/MatchedJob"
        hasCV:
          type: boolean
      required:
        - jobs
        - hasCV
...
    MatchAnalysis:
      type: object
      properties:
        jobId:
          type: string
        matchScore:
          type: integer
        strengths:
          type: array
          items:
            type: string
        gaps:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        cachedAt:
          type: string
      required:
        - jobId
        - matchScore
        - strengths
        - gaps
        - suggestions
        - cachedAt
...
    SyncResult:
      type: object
      properties:
        inserted:
          type: integer
        updated:
          type: integer
        deleted:
          type: integer
        errors:
          type: integer
        message:
          type: string
      required:
        - inserted
        - updated
        - deleted
        - errors
        - message